refactor(bloc): internal reorganization (#3011)
