+++ packages/bloc/lib/src/bloc.dart
@@ -1,164 +1,4 @@
-import 'dart:async';
-
-import 'package:bloc/bloc.dart';
-import 'package:meta/meta.dart';
-
-const _asyncRunZoned = runZoned;
-
-/// This class facilitates overriding [BlocObserver] and [EventTransformer].
-/// It should be extended by another class in client code with overrides
-/// that construct a custom implementation. The implementation in this class
-/// defaults to the base [blocObserver] and [eventTransformer] implementation.
-/// For example:
-///
-/// ```dart
-/// class MyBlocObserver extends BlocObserver {
-///   ...
-///   // A custom BlocObserver implementation.
-///   ...
-/// }
-///
-/// void main() {
-///   BlocOverrides.runZoned(() {
-///     ...
-///     // Bloc instances will use MyBlocObserver instead of the default BlocObserver.
-///     ...
-///   }, blocObserver: MyBlocObserver());
-/// }
-/// ```
-abstract class BlocOverrides {
-  static final _token = Object();
-
-  /// Returns the current [BlocOverrides] instance.
-  ///
-  /// This will return `null` if the current [Zone] does not contain
-  /// any [BlocOverrides].
-  ///
-  /// See also:
-  /// * [BlocOverrides.runZoned] to provide [BlocOverrides] in a fresh [Zone].
-  ///
-  static BlocOverrides? get current => Zone.current[_token] as BlocOverrides?;
-
-  /// Runs [body] in a fresh [Zone] using the provided overrides.
-  static R runZoned<R>(
-    R Function() body, {
-    BlocObserver? blocObserver,
-    EventTransformer? eventTransformer,
-  }) {
-    final overrides = _BlocOverridesScope(blocObserver, eventTransformer);
-    return _asyncRunZoned(body, zoneValues: {_token: overrides});
-  }
-
-  /// The [BlocObserver] that will be used within the current [Zone].
-  ///
-  /// By default, a base [BlocObserver] implementation is used.
-  BlocObserver get blocObserver => _defaultBlocObserver;
-
-  /// The [EventTransformer] that will be used within the current [Zone].
-  ///
-  /// By default, all events are processed concurrently.
-  ///
-  /// If a custom transformer is specified for a particular event handler,
-  /// it will take precendence over the global transformer.
-  ///
-  /// See also:
-  ///
-  /// * [package:bloc_concurrency](https://pub.dev/packages/bloc_concurrency) for an
-  /// opinionated set of event transformers.
-  ///
-  EventTransformer get eventTransformer => _defaultEventTransformer;
-}
-
-class _BlocOverridesScope extends BlocOverrides {
-  _BlocOverridesScope(this._blocObserver, this._eventTransformer);
-
-  final BlocOverrides? _previous = BlocOverrides.current;
-  final BlocObserver? _blocObserver;
-  final EventTransformer? _eventTransformer;
-
-  @override
-  BlocObserver get blocObserver {
-    final blocObserver = _blocObserver;
-    if (blocObserver != null) return blocObserver;
-
-    final previous = _previous;
-    if (previous != null) return previous.blocObserver;
-
-    return super.blocObserver;
-  }
-
-  @override
-  EventTransformer get eventTransformer {
-    final eventTransformer = _eventTransformer;
-    if (eventTransformer != null) return eventTransformer;
-
-    final previous = _previous;
-    if (previous != null) return previous.eventTransformer;
-
-    return super.eventTransformer;
-  }
-}
-
-/// {@template emitter}
-/// An [Emitter] is a class which is capable of emitting new states.
-///
-/// See also:
-///
-/// * [EventHandler] which has access to an [Emitter].
-///
-/// {@endtemplate}
-abstract class Emitter<State> {
-  /// Subscribes to the provided [stream] and invokes the [onData] callback
-  /// when the [stream] emits new data.
-  ///
-  /// [onEach] completes when the event handler is cancelled or when
-  /// the provided [stream] has ended.
-  ///
-  /// If [onError] is omitted, any errors on this [stream]
-  /// are considered unhandled, and will be thrown by [onEach].
-  /// As a result, the internal subscription to the [stream] will be canceled.
-  ///
-  /// If [onError] is provided, any errors on this [stream] will be passed on to
-  /// [onError] and will not result in unhandled exceptions or cancelations to
-  /// the internal stream subscription.
-  ///
-  /// **Note**: The stack trace argument may be [StackTrace.empty]
-  /// if the [stream] received an error without a stack trace.
-  Future<void> onEach<T>(
-    Stream<T> stream, {
-    required void Function(T data) onData,
-    void Function(Object error, StackTrace stackTrace)? onError,
-  });
-
-  // Subscribes to the provided [stream] and invokes the [onData] callback
-  /// when the [stream] emits new data and the result of [onData] is emitted.
-  ///
-  /// [forEach] completes when the event handler is cancelled or when
-  /// the provided [stream] has ended.
-  ///
-  /// If [onError] is omitted, any errors on this [stream]
-  /// are considered unhandled, and will be thrown by [forEach].
-  /// As a result, the internal subscription to the [stream] will be canceled.
-  ///
-  /// If [onError] is provided, any errors on this [stream] will be passed on to
-  /// [onError] and will not result in unhandled exceptions or cancelations to
-  /// the internal stream subscription.
-  ///
-  /// **Note**: The stack trace argument may be [StackTrace.empty]
-  /// if the [stream] received an error without a stack trace.
-  Future<void> forEach<T>(
-    Stream<T> stream, {
-    required State Function(T data) onData,
-    State Function(Object error, StackTrace stackTrace)? onError,
-  });
-
-  /// Whether the [EventHandler] associated with this [Emitter]
-  /// has been completed or canceled.
-  bool get isDone;
-
-  /// Emits the provided [state].
-  void call(State state);
-}
+part of bloc;
 
 /// An event handler is responsible for reacting to an incoming [Event]
 /// and can emit zero or more states via the [Emitter].
@@ -178,129 +18,6 @@ typedef EventTransformer<Event> = Stream<Event> Function(
   Stream<Event> events,
   EventMapper<Event> mapper,
 );
-
-class _Emitter<State> implements Emitter<State> {
-  _Emitter(this._emit);
-
-  final void Function(State state) _emit;
-  final _completer = Completer<void>();
-  final _disposables = <FutureOr<void> Function()>[];
-
-  var _isCanceled = false;
-  var _isCompleted = false;
-
-  @override
-  Future<void> onEach<T>(
-    Stream<T> stream, {
-    required void Function(T data) onData,
-    void Function(Object error, StackTrace stackTrace)? onError,
-  }) async {
-    final completer = Completer<void>();
-    final subscription = stream.listen(
-      onData,
-      onDone: completer.complete,
-      onError: onError ?? completer.completeError,
-      cancelOnError: onError == null,
-    );
-    _disposables.add(subscription.cancel);
-    return Future.any([future, completer.future]).whenComplete(() {
-      subscription.cancel();
-      _disposables.remove(subscription.cancel);
-    });
-  }
-
-  @override
-  Future<void> forEach<T>(
-    Stream<T> stream, {
-    required State Function(T data) onData,
-    State Function(Object error, StackTrace stackTrace)? onError,
-  }) {
-    return onEach<T>(
-      stream,
-      onData: (data) => call(onData(data)),
-      onError: onError != null
-          ? (Object error, StackTrace stackTrace) {
-              call(onError(error, stackTrace));
-            }
-          : null,
-    );
-  }
-
-  @override
-  void call(State state) {
-    assert(
-      !_isCompleted,
-      '''\n\n
-emit was called after an event handler completed normally.
-This is usually due to an unawaited future in an event handler.
-Please make sure to await all asynchronous operations with event handlers
-and use emit.isDone after asynchronous operations before calling emit() to
-ensure the event handler has not completed.
-
-  **BAD**
-  on<Event>((event, emit) {
-    future.whenComplete(() => emit(...));
-  });
-
-  **GOOD**
-  on<Event>((event, emit) async {
-    await future.whenComplete(() => emit(...));
-  });
-''',
-    );
-    if (!_isCanceled) _emit(state);
-  }
-
-  @override
-  bool get isDone => _isCanceled || _isCompleted;
-
-  void cancel() {
-    if (isDone) return;
-    _isCanceled = true;
-    _close();
-  }
-
-  void complete() {
-    if (isDone) return;
-    assert(
-      _disposables.isEmpty,
-      '''\n\n
-An event handler completed but left pending subscriptions behind.
-This is most likely due to an unawaited emit.forEach or emit.onEach. 
-Please make sure to await all asynchronous operations within event handlers.
-
-  **BAD**
-  on<Event>((event, emit) {
-    emit.forEach(...);
-  });  
-  
-  **GOOD**
-  on<Event>((event, emit) async {
-    await emit.forEach(...);
-  });
-
-  **GOOD**
-  on<Event>((event, emit) {
-    return emit.forEach(...);
-  });
-
-  **GOOD**
-  on<Event>((event, emit) => emit.forEach(...));
-
-''',
-    );
-    _isCompleted = true;
-    _close();
-  }
-
-  void _close() {
-    for (final disposable in _disposables) disposable.call();
-    _disposables.clear();
-    if (!_completer.isCompleted) _completer.complete();
-  }
-
-  Future<void> get future => _completer.future;
-}
 
 class _Handler {
   const _Handler({required this.isType, required this.type});
@@ -531,206 +248,5 @@ abstract class Bloc<Event, State> extends BlocBase<State> {
     await Future.wait<void>(_emitters.map((e) => e.future));
     await Future.wait<void>(_subscriptions.map((s) => s.cancel()));
     return super.close();
-  }
-}
-
-/// {@template cubit}
-/// A [Cubit] is similar to [Bloc] but has no notion of events
-/// and relies on methods to [emit] new states.
-///
-/// Every [Cubit] requires an initial state which will be the
-/// state of the [Cubit] before [emit] has been called.
-///
-/// The current state of a [Cubit] can be accessed via the [state] getter.
-///
-/// ```dart
-/// class CounterCubit extends Cubit<int> {
-///   CounterCubit() : super(0);
-///
-///   void increment() => emit(state + 1);
-/// }
-/// ```
-///
-/// {@endtemplate}
-abstract class Cubit<State> extends BlocBase<State> {
-  /// {@macro cubit}
-  Cubit(State initialState) : super(initialState);
-}
-
-/// {@template bloc_stream}
-/// An interface for the core functionality implemented by
-/// both [Bloc] and [Cubit].
-/// {@endtemplate}
-abstract class BlocBase<State> {
-  /// {@macro bloc_stream}
-  BlocBase(this._state) {
-    // ignore: invalid_use_of_protected_member
-    _blocObserver?.onCreate(this);
-  }
-
-  final _blocObserver = BlocOverrides.current?.blocObserver;
-
-  StreamController<State>? __stateController;
-  StreamController<State> get _stateController {
-    return __stateController ??= StreamController<State>.broadcast();
-  }
-
-  State _state;
-
-  bool _emitted = false;
-
-  /// The current [state].
-  State get state => _state;
-
-  /// The current state stream.
-  Stream<State> get stream => _stateController.stream;
-
-  /// Whether the bloc is closed.
-  ///
-  /// A bloc is considered closed once [close] is called.
-  /// Subsequent state changes cannot occur within a closed bloc.
-  bool get isClosed => _stateController.isClosed;
-
-  /// Updates the [state] to the provided [state].
-  /// [emit] does nothing if the [state] being emitted
-  /// is equal to the current [state].
-  ///
-  /// To allow for the possibility of notifying listeners of the initial state,
-  /// emitting a state which is equal to the initial state is allowed as long
-  /// as it is the first thing emitted by the instance.
-  ///
-  /// * Throws a [StateError] if the bloc is closed.
-  @protected
-  @visibleForTesting
-  void emit(State state) {
-    try {
-      if (isClosed) {
-        throw StateError('Cannot emit new states after calling close');
-      }
-      if (state == _state && _emitted) return;
-      onChange(Change<State>(currentState: this.state, nextState: state));
-      _state = state;
-      _stateController.add(_state);
-      _emitted = true;
-    } catch (error, stackTrace) {
-      onError(error, stackTrace);
-      rethrow;
-    }
-  }
-
-  /// Called whenever a [change] occurs with the given [change].
-  /// A [change] occurs when a new `state` is emitted.
-  /// [onChange] is called before the `state` of the `cubit` is updated.
-  /// [onChange] is a great spot to add logging/analytics for a specific `cubit`.
-  ///
-  /// **Note: `super.onChange` should always be called first.**
-  /// ```dart
-  /// @override
-  /// void onChange(Change change) {
-  ///   // Always call super.onChange with the current change
-  ///   super.onChange(change);
-  ///
-  ///   // Custom onChange logic goes here
-  /// }
-  /// ```
-  ///
-  /// See also:
-  ///
-  /// * [BlocObserver] for observing [Cubit] behavior globally.
-  ///
-  @mustCallSuper
-  void onChange(Change<State> change) {
-    // ignore: invalid_use_of_protected_member
-    _blocObserver?.onChange(this, change);
-  }
-
-  /// Reports an [error] which triggers [onError] with an optional [StackTrace].
-  @mustCallSuper
-  void addError(Object error, [StackTrace? stackTrace]) {
-    onError(error, stackTrace ?? StackTrace.current);
-  }
-
-  /// Called whenever an [error] occurs and notifies [BlocObserver.onError].
-  ///
-  /// **Note: `super.onError` should always be called last.**
-  ///
-  /// ```dart
-  /// @override
-  /// void onError(Object error, StackTrace stackTrace) {
-  ///   // Custom onError logic goes here
-  ///
-  ///   // Always call super.onError with the current error and stackTrace
-  ///   super.onError(error, stackTrace);
-  /// }
-  /// ```
-  @protected
-  @mustCallSuper
-  void onError(Object error, StackTrace stackTrace) {
-    // ignore: invalid_use_of_protected_member
-    _blocObserver?.onError(this, error, stackTrace);
-  }
-
-  /// Closes the instance.
-  /// This method should be called when the instance is no longer needed.
-  /// Once [close] is called, the instance can no longer be used.
-  @mustCallSuper
-  Future<void> close() async {
-    // ignore: invalid_use_of_protected_member
-    _blocObserver?.onClose(this);
-    await _stateController.close();
-  }
-}
-
-late final _defaultBlocObserver = _DefaultBlocObserver();
-late final _defaultEventTransformer = (Stream events, EventMapper mapper) {
-  return events
-      .map(mapper)
-      .transform<dynamic>(const _FlatMapStreamTransformer<dynamic>());
-};
-
-class _DefaultBlocObserver extends BlocObserver {}
-
-class _FlatMapStreamTransformer<T> extends StreamTransformerBase<Stream<T>, T> {
-  const _FlatMapStreamTransformer();
-
-  @override
-  Stream<T> bind(Stream<Stream<T>> stream) {
-    final controller = StreamController<T>.broadcast(sync: true);
-
-    controller.onListen = () {
-      final subscriptions = <StreamSubscription<dynamic>>[];
-
-      final outerSubscription = stream.listen(
-        (inner) {
-          final subscription = inner.listen(
-            controller.add,
-            onError: controller.addError,
-          );
-
-          subscription.onDone(() {
-            subscriptions.remove(subscription);
-            if (subscriptions.isEmpty) controller.close();
-          });
-
-          subscriptions.add(subscription);
-        },
-        onError: controller.addError,
-      );
-
-      outerSubscription.onDone(() {
-        subscriptions.remove(outerSubscription);
-        if (subscriptions.isEmpty) controller.close();
-      });
-
-      subscriptions.add(outerSubscription);
-
-      controller.onCancel = () {
-        if (subscriptions.isEmpty) return null;
-        final cancels = [for (final s in subscriptions) s.cancel()];
-        return Future.wait(cancels).then((_) {});
-      };
-    };
-
-    return controller.stream;
   }
 }
