+++ packages/bloc/lib/src/bloc_observer.dart
@@ -1,5 +1,4 @@
-import 'package:bloc/bloc.dart';
-import 'package:meta/meta.dart';
+part of bloc;
 
 /// An interface for observing the behavior of [Bloc] instances.
 abstract class BlocObserver {
