+++ packages/bloc/lib/src/cubit.dart
@@ -1,0 +1,24 @@
+part of bloc;
+
+/// {@template cubit}
+/// A [Cubit] is similar to [Bloc] but has no notion of events
+/// and relies on methods to [emit] new states.
+///
+/// Every [Cubit] requires an initial state which will be the
+/// state of the [Cubit] before [emit] has been called.
+///
+/// The current state of a [Cubit] can be accessed via the [state] getter.
+///
+/// ```dart
+/// class CounterCubit extends Cubit<int> {
+///   CounterCubit() : super(0);
+///
+///   void increment() => emit(state + 1);
+/// }
+/// ```
+///
+/// {@endtemplate}
+abstract class Cubit<State> extends BlocBase<State> {
+  /// {@macro cubit}
+  Cubit(State initialState) : super(initialState);
+}
